test: Refactor setupServerless so it works in Node.js v6
