+++ package.json
@@ -57,7 +57,7 @@
     "@babel/core": "^7.10.5",
     "@babel/preset-env": "^7.10.4",
     "@babel/register": "^7.10.5",
-    "@serverless/eslint-config": "^2.1.1",
+    "@serverless/eslint-config": "^2.2.0",
     "@serverless/test": "^4.3.2",
     "aws-sdk": "^2.720.0",
     "chai": "^4.2.0",
